* Added helper method for getting large buff icons * Toxic imbue now gives less toxic gas when about to wear off, but gives corrosion/ooze resists too. * Huntress starts with holster * Fixed exploit with stone (I think) * Wealth should stop dropping nothing * Chilling should work properly * Sniper shot always surprise attacks * Earthroot now roots enemies, applies regular effect to warden in addition to special * Starflower now gives adrenaline to warden, which can stack. * Some comments added

+++ core/src/main/java/com/zrp200/lustrouspixeldungeon/plants/Starflower.java
@@ -24,6 +24,7 @@ package com.zrp200.lustrouspixeldungeon.plants;
 import com.watabou.utils.Random;
 import com.zrp200.lustrouspixeldungeon.Dungeon;
 import com.zrp200.lustrouspixeldungeon.actors.Char;
+import com.zrp200.lustrouspixeldungeon.actors.buffs.Adrenaline;
 import com.zrp200.lustrouspixeldungeon.actors.buffs.Bless;
 import com.zrp200.lustrouspixeldungeon.actors.buffs.Buff;
 import com.zrp200.lustrouspixeldungeon.actors.buffs.Recharging;
@@ -39,17 +40,18 @@ public class Starflower extends Plant {
 
 	@Override
 	public void activate( Char ch ) {
-
 		if (ch != null) {
 			Buff.prolong(ch, Bless.class, Bless.DURATION);
 			if (ch instanceof Hero && ((Hero) ch).subClass == HeroSubClass.WARDEN){
-				Buff.prolong(ch, Recharging.class, Bless.DURATION);
+				Buff.affect(ch, Adrenaline.class, Adrenaline.DURATION);
 			}
 		}
+	}
 
-		if (Random.Int(5) == 0){
-			Dungeon.level.drop(new Seed(), pos).sprite.drop();
-		}
+	@Override
+	public void wither() {
+		super.wither();
+		if (Random.Int(5) == 0) Dungeon.level.drop(new Seed(), pos).sprite.drop();
 	}
 
 	public static class Seed extends Plant.Seed{
